+++ latex2mathml/commands.py
@@ -1,0 +1,12 @@
+#!/usr/bin/python
+
+__author__ = 'Ronie Martinez'
+
+MATRICES = (r'\matrix', r'\matrix*',
+            r'\pmatrix', r'\pmatrix*',
+            r'\bmatrix', r'\bmatrix*',
+            r'\Bmatrix', r'\Bmatrix*',
+            r'\vmatrix', r'\vmatrix*',
+            r'\Vmatrix', r'\Vmatrix*')
+
+SPACES = (r'\,', r'\:', r'\;', '\\', r'\quad', r'\qquad')